#111 Ask for confirmation and give feedback about the success/failure of Register sequence#113 Ask for confirmation and give feedback about the success/failure of Remove sequence

+++ client/src/app/components/sequences/sequences.component.ts
@@ -10,6 +10,8 @@ import { Child } from '../../models/child';
 import { ChildrenService } from '../../services/children/children.service';
 import { SequencesService } from '../../services/sequences/sequences.service';
 import { PaginationService } from '../../services/pagination/pagination.service';
+import { Overlay } from 'ngx-modialog';
+import { Modal } from 'ngx-modialog/plugins/bootstrap';
 
 @Component({
   selector: 'app-sequences',
@@ -20,10 +22,15 @@ export class SequencesComponent implements OnInit, OnChanges {
 
   public childId = 0;
   public sequenceId = 0;
+  public sequenceName;
   public child;
   public sequences: Array<Sequence>;
-  public loading: boolean = false;
+  public loading = false;
   public children: Array<Child>;
+  public success = false;
+  public failure = false;
+  public textSuccess;
+  public textFailure;
 
   // pager object
   pager: any = {};
@@ -32,7 +39,7 @@ export class SequencesComponent implements OnInit, OnChanges {
   pagedItems: any[];
 
   constructor(public route: ActivatedRoute, public childrenService: ChildrenService, public sequencesService: SequencesService,
-    public paginationService: PaginationService, public router: Router, public location: Location) { }
+    public paginationService: PaginationService, public router: Router, public location: Location, public modal: Modal) { }
 
   ngOnInit() {
     this.fetchSequences();
@@ -44,6 +51,7 @@ export class SequencesComponent implements OnInit, OnChanges {
 
   public fetchSequences() {
      this.loading = true;
+     this.updateSuccessFailure();
      this.route.params
       .switchMap((params: Params) => Observable.of(params))
       .subscribe(params => {
@@ -118,14 +126,31 @@ export class SequencesComponent implements OnInit, OnChanges {
   }
 
   public deleteSequence(sequenceId) {
-    this.loading = true;
-    this.sequencesService.deleteSequence(sequenceId).subscribe(
-      result => this.fetchSequences(),
-      err => {
-        console.log('Error deleting sequence');
-        this.loading = false;
+
+    const dialogRef = this.modal.confirm().size('lg').isBlocking(true).showClose(false).okBtn('Sim').cancelBtn('Não')
+      .title('Eliminar aula').body('Tem a certeza que pretende eliminar a aula?').open();
+
+    dialogRef.then(dialogRef => { dialogRef.result.then(result => {
+      if (result) {
+        this.sequencesService.deleteSequence(sequenceId).subscribe(
+          res => {
+            this.sequencesService.setSuccess(true);
+              this.sequencesService.setFailure(false);
+              this.sequencesService.setTextSuccess('Aula eliminada com sucesso.');
+              this.fetchSequences();
+          },
+          err => {
+            console.log('Error deleting sequence');
+            this.sequencesService.setSuccess(false);
+            this.sequencesService.setFailure(true);
+            this.sequencesService.setTextFailure('Não foi possível eliminar a aula.');
+            this.updateSuccessFailure();
+          }
+        );
+      } else {
+        this.goBack();
       }
-    )
+    }).catch(() => {})});
   }
 
   public truncate(str: String, size: number) {
@@ -144,6 +169,13 @@ export class SequencesComponent implements OnInit, OnChanges {
     this.location.back();
   }
 
+  public updateSuccessFailure() {
+    this.success = this.sequencesService.getSuccess();
+    this.failure = this.sequencesService.getFailure();
+    this.textSuccess = this.sequencesService.getTextSuccess();
+    this.textFailure = this.sequencesService.getTextFailure();
+  }
+
   setPage(page: number) {
     if (page < 1 || page > this.pager.totalPages) {
       return;
@@ -155,4 +187,12 @@ export class SequencesComponent implements OnInit, OnChanges {
     // get current page of items
     this.pagedItems = this.sequences.slice(this.pager.startIndex, this.pager.endIndex + 1);
     }
+
+    reset() {
+      this.sequencesService.success = false;
+      this.sequencesService.failure = false;
+      this.sequencesService.textFailure = '';
+      this.sequencesService.textSuccess = '';
+    }
+
 }
